pyproject.toml
[project]
name = "attention-visualization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.22.4",
    "matplotlib>=3.10.8",
    "pillow>=12.2.0",
    "pycocotools>=2.0.11",
    "requests>=2.33.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0",
]
